feat: extracted list base component
